refactor(gateway): extract message_handler.py from gateway.py — split step 5/6
Schritt 5 von 6 im Gateway-Split-Plan.

What moved
- 6 message-flow helpers extracted into `gateway/message_handler.py`:
    * `handle_message(gw, msg, stream_callback)` — public API entry point
      orchestrating the full message lifecycle
    * `resolve_agent_route(gw, msg)` — Phase 1: agent routing, session,
      working memory, skills, workspace
    * `prepare_execution_context(gw, msg, session, wm, route)` — Phase 2:
      profiler, budget, run-recorder, policy snapshot
    * `make_status_callback(gw, channel_name, session_id)` — fire-and-forget
      callback factory for live UI status updates
    * `make_pipeline_callback(gw, channel_name, session_id)` — same for
      structured PGE pipeline events
    * `formulate_response(gw, msg_text, results, wm, stream_callback)` —
      observer-aware response synthesis with streaming-path fallback

`gateway.py`: 4 249 -> 3 399 lines (-20 % in this PR; cumulative -49 %
from the pre-split 6 628 LOC after PRs 1-5).

Pre-extraction grep (Lesson from PRs 1-4)
- Class-level decorators / constants in range [2269..3185]: only the
  `@staticmethod` for `_is_cu_plan` wrapper (already in pge_loop split).
  No additional `ClassVar` constants needed restoration this time.

Imports + missing references
- Stdlib: `asyncio`, `time` (others were unused after extraction).
- Cross-module runtime imports: `RouteDecision`, `AutonomousOrchestrator`,
  `run_pge_with_observer_directive`, `_classify_attachments`,
  `_build_video_attachment`, `_extract_uuid_from_path` (top-level helpers
  in gateway.py — imported back).
- `cognithor.models`: `ActionPlan`, `AgentResult`, `AuditEntry`,
  `ComplianceViolation` (from `security.compliance_engine`),
  `Message`, `MessageRole`, `IncomingMessage`, `OutgoingMessage`,
  `ToolResult`, `WorkingMemory`.
- `cognithor.i18n.t` (used pervasively for translated strings).
- `RouteDecision` import in gateway.py needed `# noqa: TC001` because
  Pydantic + the wrapper signature use it at runtime.

Test adapter (5th source-string-probe occurrence in this split)
- `tests/test_gateway/test_gateway_video_wiring.py::test_gateway_wraps_build_video_attachment_in_to_thread`
  scanned `gateway.py` source for `asyncio.to_thread(_build_video_attachment`.
  After this PR the call lives in `message_handler.py`. Updated to
  concatenate both sources before scanning (same pattern as
  `test_audit_fixes`, `test_video_cleanup_session_wiring`,
  `test_inv4_budget_hard_ceilings`).

Wrapper-signature byte-exactness
- All 6 wrapper signatures copied verbatim from main pre-split. One
  `self.` reference inside `handle_message` body that the regex missed
  (inside `hasattr(self, "_autonomous_orchestrator")`) caught by F821
  and fixed manually.

Verification
- `pytest tests/test_gateway/ tests/test_core/` -> **2 704 passed, 1 skipped**.
- `ruff check + format` -> All checks passed.
- `python -c "from cognithor.gateway.gateway import Gateway; from cognithor.gateway import message_handler"` -> OK.

Plan reference: project_v0960_refactor_backlog.md "gateway.py split"
+ inline architect blueprint. Next + final PR: `lifecycle.py` (PR 6/6)
extracts `start()`, `shutdown()`, vLLM hooks.
